+++ test/components/bullet/bullet-api.func-spec.js
@@ -37,6 +37,31 @@ const dataset2 = [{
   markerColors: ['#000000']
 }];
 
+const dataset3 = [{
+  data: [{
+    title: 'Revenue (New)',
+    subtitle: 'DKK$, in thousands',
+    ranges: [1500000, 2250000, 3000000, 4000000, 6000000],
+    measures: [2200000, 2700000],
+    markers: [2500000]
+  }],
+  barColors: ['#C0EDE3', '#8ED1C6', '#69ADA3', '#448D83', '#206B62'],
+  lineColors: ['#000000', '#000000', '#000000'],
+  markerColors: ['#000000']
+}];
+
+function formatToUnits(num, digits) {
+  const units = ['k', 'M', 'G', 'T', 'P', 'E', 'Z', 'Y'];
+  let decimal;
+  for (let i = units.length - 1; i >= 0; i--) {
+    decimal = Math.pow(1000, (i + 1));
+    if (num <= -decimal || num >= decimal) {
+      return +(num / decimal).toFixed(digits) + units[i];
+    }
+  }
+  return num;
+}
+
 describe('Bullet Chart API', () => {
   beforeEach((done) => {
     bulletEl = null;
@@ -87,5 +112,83 @@ describe('Bullet Chart API', () => {
     triggerContextmenu(document.body.querySelector('.range'));
 
     expect(spyEvent).toHaveBeenTriggered();
+  });
+
+  it('Should format values with d3 formatter string by Dataset', (done) => {
+    const ds = JSON.parse(JSON.stringify(dataset3));
+    ds[0].format = {
+      ranges: '.1s',
+      difference: '.1s'
+    };
+    bulletObj.destroy();
+    bulletObj = new Bullet(bulletEl, { dataset: ds, animate: false });
+
+    setTimeout(() => {
+      expect(document.body.querySelectorAll('.tick').length).toEqual(7);
+      expect(document.body.querySelectorAll('.tick')[0].textContent).toEqual('0');
+      expect(document.body.querySelectorAll('.tick')[6].textContent).toEqual('6M');
+      expect(document.body.querySelector('.inverse').textContent).toEqual('200k');
+      done();
+    }, 300);
+  });
+
+  it('Should format values with d3 formatter string by Settings', (done) => {
+    const ds = JSON.parse(JSON.stringify(dataset3));
+    bulletObj.destroy();
+    bulletObj = new Bullet(bulletEl, {
+      dataset: ds,
+      animate: false,
+      format: {
+        ranges: '.1s',
+        difference: '.1s'
+      }
+    });
+
+    setTimeout(() => {
+      expect(document.body.querySelectorAll('.tick').length).toEqual(7);
+      expect(document.body.querySelectorAll('.tick')[0].textContent).toEqual('0');
+      expect(document.body.querySelectorAll('.tick')[6].textContent).toEqual('6M');
+      expect(document.body.querySelector('.inverse').textContent).toEqual('200k');
+      done();
+    }, 300);
+  });
+
+  it('Should format values with callback function by Dataset', (done) => {
+    const ds = JSON.parse(JSON.stringify(dataset3));
+    ds[0].format = {
+      ranges: d => formatToUnits(d),
+      difference: d => formatToUnits(d)
+    };
+    bulletObj.destroy();
+    bulletObj = new Bullet(bulletEl, { dataset: ds, animate: false });
+
+    setTimeout(() => {
+      expect(document.body.querySelectorAll('.tick').length).toEqual(7);
+      expect(document.body.querySelectorAll('.tick')[0].textContent).toEqual('0');
+      expect(document.body.querySelectorAll('.tick')[6].textContent).toEqual('6M');
+      expect(document.body.querySelector('.inverse').textContent).toEqual('200k');
+      done();
+    }, 300);
+  });
+
+  it('Should format values with callback function by Settings', (done) => {
+    const ds = JSON.parse(JSON.stringify(dataset3));
+    bulletObj.destroy();
+    bulletObj = new Bullet(bulletEl, {
+      dataset: ds,
+      animate: false,
+      format: {
+        ranges: d => formatToUnits(d),
+        difference: d => formatToUnits(d)
+      }
+    });
+
+    setTimeout(() => {
+      expect(document.body.querySelectorAll('.tick').length).toEqual(7);
+      expect(document.body.querySelectorAll('.tick')[0].textContent).toEqual('0');
+      expect(document.body.querySelectorAll('.tick')[6].textContent).toEqual('6M');
+      expect(document.body.querySelector('.inverse').textContent).toEqual('200k');
+      done();
+    }, 300);
   });
 });